requirements.txt
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
enum34==1.1.6
idna==2.6
lxml==4.1.1
namedlist==1.7
pluggy==0.6.0
py==1.5.2
pytest==3.3.0
pytest-allure-adaptor==1.7.9
requests==2.18.4
six==1.11.0
urllib3==1.22
